+++ react-app/src/components/ChartTest/index.js
@@ -2,7 +2,6 @@ import '../../stylesheets/chart.css'
 // import ChartDrawing from '../ChartDrawing'
 // import Search from '../Search'
 // import { useState } from 'react'
-import RenderChart from '../WatchList/RenderChart'
 import PlaceHolder from '../PlaceHolder'
 
 const ChartTest = () => {
@@ -11,9 +10,6 @@ const ChartTest = () => {
   return (
     <>
     <PlaceHolder />
-    <div style={{padding: '1rem'}}>
-      {samples.map(ticker => <RenderChart key={ticker} symbol={ticker} />)}
-    </div>
     </>
   )
 }
