+++ react-app/src/components/PlaceHolder/index.js
@@ -1,8 +1,7 @@
 import styles from './placeholder.module.css'
 
 const PlaceHolder = ({ height }) => {
-    console.log(height)
-    return <div className={styles.placeholder} style={{height: `${height || 325}px`}}><i class="fa-solid fa-spinner"></i></div>
+    return <div className={styles.placeholder} style={{height: `${height || 329}px`}}><i className={`fa-solid fa-spinner ${styles.loading}`}></i></div>
 }
 
 export default PlaceHolder
