+++ react-app/src/components/PlaceHolder/placeholder.module.css
@@ -1,7 +1,24 @@
 .placeholder {
     width: 100%;
-    background-color: #e2e2e2;
+    background-color: rgb(227, 227, 227, 0.3);
     display: flex;
     align-items: center;
     justify-content: center;
+    border-radius: 4px;
+}
+
+.loading {
+    aspect-ratio: 1/1;
+    animation: loader 1s ease infinite;
+    font-size: 2rem;
+}
+
+@keyframes loader {
+    0% {
+        transform: rotate(0deg);
+    }
+
+    100% {
+        transform: rotate(360deg);
+    }
 }
