+++ react-app/src/components/SmallChart/render.module.css
