+++ react-app/src/components/WatchList/index.js
@@ -4,11 +4,10 @@ import { useEffect, useState } from 'react';
 import NewWatchList from './watchlist_form';
 import './index.css';
 import UpdateButton from './Update/UpdateButton';
-// import SmallChart from '../SmallChart';
+import SmallChart from '../SmallChart';
 import StockPrice from './StockPrice';
 import { Modal } from "../Modals/Modal";
 import UpdateWatchlistForm from "./Update/watchlist_updateForm";
-import RenderChart from './RenderChart';
 
 const WatchList = () => {
     const dispatch = useDispatch();
@@ -91,7 +90,7 @@ const WatchList = () => {
                             {openings[i] &&
                                 <div className='watchlist-stocks-container'>
                                     {watchlist.watchlist_stocks.length > 0 &&
-                                        watchlist.watchlist_stocks.map(stock => (<RenderChart symbol={stock.stock_symbol} />))
+                                        watchlist.watchlist_stocks.map(stock => (<SmallChart symbol={stock.stock_symbol} />))
                                     }
                                 </div>
                             }
